Implements wildcard input modification
Adds functionality to modify the input string, handling wildcards
within it. It tokenizes the input and expands wildcard patterns
before reconstructing the modified input. Includes a debugging
print function to display the state of modified input.
not completed yet, it needs a fucntion that is in the scope of t_split branch. which ıll go modify right now.