Updates project maintenance and signal handling
Updates the Makefile and source files to reflect the change of maintainer.

Replaces `printf` with `ft_putstr_fd` in the echo built-in to use file descriptors for output, likely for redirection purposes.

Changes the "Quit: 3" message to "Quit (core dumped)" for SIGQUIT signals to provide more clarity.

Adds a suppression file for Valgrind to ignore known leaks.